+++ src/index.ts
@@ -4,3 +4,4 @@ export * from './index-to-name/index-to-name';
 export * from './add-subtract-time/add-subtract-time';
 export * from './is-valid-date/is-valid-date';
 export * from './is-leap-year/is-leap-year';
+export * from './get-next-occurrence/get-next-occurrence';